+++ assembler/src/include/mph_index/mph_index.h
@@ -73,7 +73,7 @@ class MPHIndex {
   uint32_t perfect_hash(const Key& x) const;  // way faster than the minimal
   template <class SeededHashFcn, class Key>  // must agree with Reset
   uint32_t perfect_square(const Key& x) const;  // even faster but needs square=true
-  uint32_t minimal_perfect_hash_size() const { return (uint32_t) size(); }
+  uint32_t minimal_perfect_hash_size() const { return size(); }
   template <class SeededHashFcn, class Key>  // must agree with Reset
   uint32_t minimal_perfect_hash(const Key& x) const;
